game_v2/requirements.txt
# Core dependencies
pyyaml>=6.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Scripting
lupa>=2.0

# LLM providers
openai>=1.0.0
google-generativeai>=0.3.0
generative-ai-hub-sdk[all]>=4.12.0

# TTS (Cloud)
google-cloud-texttospeech>=2.14.0

# Audio output
pyaudio>=0.2.11
pygame>=2.5.0
